+++ src/main/java/com/spring/sns/controller/UserController.java
@@ -1,7 +1,7 @@
 package com.spring.sns.controller;
 
-import com.spring.sns.dto.UserCheckResponseDto;
-import com.spring.sns.dto.UserEditResponseDto;
+import com.spring.sns.dto.users.UserCheckResponseDto;
+import com.spring.sns.dto.users.UserEditResponseDto;
 import com.spring.sns.service.UserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
