+++ src/main/java/com/spring/sns/dto/comments/CommentCreateResponseDto.java
@@ -1,15 +1,18 @@
-package com.spring.sns.dto;
+package com.spring.sns.dto.comments;
 
-public class CommentDeleteResponseDto {
+public class CommentCreateResponseDto {
 
+    // 속성
     private int status;
     private String message;
 
-    public CommentDeleteResponseDto(int status, String message) {
+    // 생성자
+    public CommentCreateResponseDto(int status, String message) {
         this.status = status;
         this.message = message;
     }
 
+    // 기능
     public int getStatus() {
         return status;
     }
