+++ src/main/java/com/spring/sns/dto/comments/CommentDeleteResponseDto.java
@@ -1,18 +1,15 @@
-package com.spring.sns.dto;
+package com.spring.sns.dto.comments;
 
-public class CommentCreateResponseDto {
+public class CommentDeleteResponseDto {
 
-    // 속성
     private int status;
     private String message;
 
-    // 생성자
-    public CommentCreateResponseDto(int status, String message) {
+    public CommentDeleteResponseDto(int status, String message) {
         this.status = status;
         this.message = message;
     }
 
-    // 기능
     public int getStatus() {
         return status;
     }
