+++ src/main/java/com/spring/sns/dto/users/UserCheckRequestDto.java
@@ -1,4 +1,4 @@
-package com.spring.sns.dto;
+package com.spring.sns.dto.users;
 
 public class UserCheckRequestDto {
     private Long id;
