+++ src/main/java/com/spring/sns/dto/users/UserEditRequestDto.java
@@ -1,4 +1,4 @@
-package com.spring.sns.dto;
+package com.spring.sns.dto.users;
 
 public class UserEditRequestDto {
     //속성
